Extract swatch presets logic from LightCard to new Swatches component
